README
Java code to change dollar amounts (in cents) to quarters, dimes, nickels and pennies. 

Enter Dollar amount (in cents) for the money you want to change. The code will give you details about the change that the machine should provide to you. For example 2 dollars and 35 cents is entered as 235.